Silence a format warning with cdparanoia.
Although cdparanoia returns a long value, it's just a cast of an int,
while libcdio uses uint8_t (which both is sufficient for the possible
99 tracks).


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@36531 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ stream/stream_cdda.c
@@ -374,7 +374,7 @@ static int open_cdda(stream_t *st,int m, void* opts, int* file_format) {
   }
 
   cd_info = cd_info_new();
-  mp_msg(MSGT_OPEN,MSGL_INFO,MSGTR_MPDEMUX_CDDA_AudioCDFoundWithNTracks,cdda_tracks(cdd));
+  mp_msg(MSGT_OPEN,MSGL_INFO,MSGTR_MPDEMUX_CDDA_AudioCDFoundWithNTracks,(int)cdda_tracks(cdd));
   for(i=0;i<cdd->tracks;i++) {
 	  char track_name[80];
 	  long sec=cdda_track_firstsector(cdd,i+1);
